requirements.txt
fastapi
boto3
pillow
python-multipart
pytest
moto
uvicorn